init: start authentification & jwt task
Update function, change code
