+++ .eslintrc.json
@@ -41,6 +41,7 @@
     "node/no-unsupported-features/es-syntax": "off",
     "no-underscore-dangle": "off",
     "import/prefer-default-export": "off",
+    "import/no-cycle": "off",
     "node/no-missing-import": [
       "error",
       { "tryExtensions": [".js", ".ts", ".json"] }
