+++ src/resources/boards/board.repository.ts
@@ -1,6 +1,6 @@
 import { getConnectionToDatabase } from '../../common/database';
 import { Board } from './board.model';
-import { IBoard } from './board.types';
+import { IBoard } from '../types';
 
 const repository = getConnectionToDatabase()!.getRepository(Board);
 
@@ -11,13 +11,13 @@ export const get = async (id: string): Promise<IBoard | undefined> => repository
 export const create = async (board: IBoard): Promise<IBoard> => repository.save(board)
 
 export const update = async (id: string, updateBody: Partial<IBoard>): Promise<IBoard> => {
-  const { columns, ...otherBody } = updateBody
-  await repository.update(id, otherBody)
-  const board = await get(id)
+  const { columns, ...otherBody } = updateBody;
+  await repository.update(id, otherBody);
+  const board = await get(id);
   return board!
 }
 
 export const delBoard = async (id: string): Promise<boolean> => {
-  const res = await repository.delete(id)
+  const res = await repository.delete(id);
   return !!res.affected
 };
