--- src/resources/boards/board.types.ts
@@ -1,9 +1,0 @@
-import { IColumn } from './column.types';
-
-interface IBoard {
-  id: string;
-  title: string;
-  columns: IColumn[];
-}
-
-export { IBoard };
