--- src/resources/boards/column.types.ts
@@ -1,7 +1,0 @@
-interface IColumn {
-  id: string;
-  title: string;
-  order: number;
-}
-
- export { IColumn };
