+++ src/resources/tasks/task.model.ts
@@ -1,9 +1,9 @@
 import { v4 as uuid } from 'uuid';
 import { Entity, Column, PrimaryGeneratedColumn, ManyToOne } from 'typeorm';
-import { ITask } from './task.types';
 import { User } from "../users/user.model";
 import { Board } from "../boards/board.model";
 import { BoardColumn } from "../boards/column.model";
+import { ITask } from '../types';
 
 @Entity()
 class Task implements ITask {
