--- src/resources/tasks/task.types.ts
@@ -1,11 +1,0 @@
-interface ITask {
-  id: string;
-  title: string;
-  order: number;
-  description: string;
-  boardId: string;
-  columnId: string | null;
-  userId: string | null;
-}
-
-export { ITask };
