--- src/resources/users/user.types.ts
@@ -1,8 +1,0 @@
-interface IUser {
-  id: string;
-  name: string;
-  login: string;
-  password: string;
-}
-
-export { IUser };
